requirements/gryphon+gds.txt
alchimia==0.8.1
alembic==0.6.0
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
Babel==2.6.0
backports.shutil-get-terminal-size==1.0.0
cement==2.10.12
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coinbase==1.0.4
colorama==0.4.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.4.2
Cython==0.20.1
decorator==4.3.0
Delorean==1.0.0
enum34==1.1.6
futures==3.2.0
gryphon-cdecimal==2.3
gryphon-money==1.2.1
gryphon-pusherclient==0.2.1
humanize==0.5.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
ipython==5.7.0
ipython-genutils==0.2.0
line-profiler==2.1.2
Mako==1.0.7
MarkupSafe==1.0
mock==1.0.1
more-itertools==4.3.0
MySQL-python==1.2.5
nose==1.3.7
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pika==0.9.14
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
pylibmc==1.6.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-dotenv==0.8.2
pytz==2018.5
raven==6.9.0
redis==2.10.6
rednose==1.3.0
requests==2.19.1
requests-futures==0.9.7
requests-toolbelt==0.8.0
retrying==1.3.3
scandir==1.7
service-identity==18.1.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.14
sure==1.2.9
termcolor==1.1.0
termstyle==0.1.11
traitlets==4.3.2
treq==18.6.0
Twisted==17.9.0
txaio==18.8.1
txredis==2.4
tzlocal==1.5.1
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.48.0
zope.interface==4.6.0
